ARD
This Arduino Adapter is an innovative device designed to function as a bridge between a vehicle's OBD-II data and Arduino (compatible with all embedded platforms). It comes equipped with an open-source Arduino library for effortless integration. Alongside granting convenient access to OBD-II data, the adapter incorporates a 9-DOF motion sensor and voltmeter to measure the vehicle's battery power. Powered directly from the OBD-II port, it efficiently converts the received power to 5V, providing a reliable power source for connected devices.